Merged branch RIP into Crexzyz. A duplicate function generated in the merge operation was deleted.

+++ AppIntegrador/Models/Responde_respuesta_con_opciones.cs
@@ -23,9 +23,9 @@ namespace AppIntegrador.Models
         public string FCodigo { get; set; }
         public string Username { get; set; }
         public string CSigla { get; set; }
-        public int GNumero { get; set; }
+        public byte GNumero { get; set; }
         public int GAnno { get; set; }
-        public int GSemestre { get; set; }
+        public byte GSemestre { get; set; }
         public System.DateTime Fecha { get; set; }
         public string PCodigo { get; set; }
         public string Justificacion { get; set; }
